META-340: Unified usages of the term "object track" in the Server and the Client codebase
<==== Mercurial data ====>
Branch: meta_4.0

+++ vms/libs/common/unit_tests/src/network/tcp_connection_processor_ut.cpp
@@ -49,7 +49,7 @@ protected:
         static const int kSteps = 64;
         std::vector<char> buffer(kTotalTestBytes / kSteps);
         for (int i = 0; i < kSteps; ++i)
-            ASSERT_TRUE(sendData(buffer.data(), buffer.size()));
+            ASSERT_TRUE(sendData(buffer.data(), (int) buffer.size()));
     }
 
 };
